1、针对EGUI_VIEW_API_DEFINE做了keil_v5编译器适配，通过定义EGUI_CONFIG_AC5宏来进行开启.目前修改涉及到egui_view_image_init与egui_view_progress_bar_init 2、修改了针对将-1给无符号整形数进行赋值及比较的代码，进行确定的类型转换; 3、以上修改只是在AC5上可以编译通过，并没有进行其它测试;

+++ src/anim/egui_animation.c
@@ -124,7 +124,7 @@ void egui_animation_update(egui_animation_t *self, uint32_t current_time)
         egui_animation_notify_start(self);
     }
 
-    if (self->start_time == -1)
+    if (self->start_time == (uint32_t)-1)
     {
         self->start_time = current_time;
     }
@@ -181,7 +181,7 @@ void egui_animation_update(egui_animation_t *self, uint32_t current_time)
             }
 
             // Restart animation
-            self->start_time = -1;
+            self->start_time = (uint32_t)-1;
             egui_animation_notify_repeat(self);
         }
     }
